clean up inventory route
