+++ app/api/inventory/route.ts
@@ -246,75 +246,3 @@ export async function POST(request: Request) {
     return NextResponse.json({ error: 'Failed to create inventory item' }, { status: 500 });
   }
 }
-
-/**
- * PATCH /api/inventory/[id]
- * Updates an existing inventory item.
- */
-export async function PATCH(
-  request: NextRequest,
-  { params }: { params: { id: string } }
-) {
-  try {
-    const itemId = params.id;
-    const body = await request.json();
-    
-    // Allow partial updates, so make all fields optional
-    const updateSchema = createInventoryItemSchema.partial();
-    const validation = updateSchema.safeParse(body);
-
-    if (!validation.success) {
-      return NextResponse.json({ error: 'Invalid input', details: validation.error.flatten() }, { status: 400 });
-    }
-
-    const { purchasedAt, ...data } = validation.data;
-    const updateData: any = { ...data };
-    
-    if (purchasedAt) {
-      updateData.purchasedAt = new Date(purchasedAt);
-    }
-
-    const updatedItem = await prisma.inventoryItem.update({
-      where: { id: itemId },
-      data: updateData,
-    });
-
-    return NextResponse.json(updatedItem);
-  } catch (error) {
-    console.error('Failed to update inventory item:', error);
-    if (error instanceof Error && 'code' in error) {
-      if ((error as any).code === 'P2025') {
-        return NextResponse.json({ error: 'Inventory item not found' }, { status: 404 });
-      }
-      if ((error as any).code === 'P2002') {
-        return NextResponse.json({ error: 'An inventory item with this SKU already exists' }, { status: 409 });
-      }
-    }
-    return NextResponse.json({ error: 'Failed to update inventory item' }, { status: 500 });
-  }
-}
-
-/**
- * DELETE /api/inventory/[id]
- * Deletes an inventory item.
- */
-export async function DELETE(
-  request: NextRequest,
-  { params }: { params: { id: string } }
-) {
-  try {
-    const itemId = params.id;
-
-    await prisma.inventoryItem.delete({
-      where: { id: itemId },
-    });
-
-    return new NextResponse(null, { status: 204 });
-  } catch (error) {
-    console.error('Failed to delete inventory item:', error);
-    if (error instanceof Error && 'code' in error && (error as any).code === 'P2025') {
-      return NextResponse.json({ error: 'Inventory item not found' }, { status: 404 });
-    }
-    return NextResponse.json({ error: 'Failed to delete inventory item' }, { status: 500 });
-  }
-}
